// src/frontend/webapp/modilist/src/services/swagger/api/models/sales-order-line-item-state.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ModilistAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @enum {string}
 */
export enum SalesOrderLineItemState {
    None = 'None',
    Sold = 'Sold',
    Returned = 'Returned'
}